% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condense.R
\name{condense}
\alias{condense}
\alias{condense_alltypex}
\title{Condense Levels}
\usage{
condense_alltypex(x)
}
\arguments{
\item{x}{vector. Data to be converted.}
}
\description{
Data, potentially already grouped, can be split into fewer groups with
these methods.
}
\section{Functions}{
\itemize{
\item \code{condense_alltypex}: Convert dwelling type from type/age to just type.
}}

