% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{count_by_group}
\alias{count_by_group}
\title{Summarise Proportions by Group}
\usage{
count_by_group(data, prop, ...)
}
\arguments{
\item{data}{tibble.}

\item{prop}{unquoted expression. Column to count.}

\item{...}{passed to \code{\link[dplyr]{group_by}}. Columns to group by such
that the sum of \code{prop} over each of these groups will be 1.}
}
\description{
Get counts and proportions of a factor by group level specified in \code{...}
}
\examples{
\dontrun{
  count_by_group(mtcars, prop = gear, cyl, carb)
}

}
