% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-lm.R
\name{custom_tidy_tbl}
\alias{custom_tidy_tbl}
\title{Custom Tidy.lm Table}
\usage{
custom_tidy_tbl(fit, digits)
}
\arguments{
\item{fit}{\code{lm} object.}

\item{digits}{integer. Significant figures to round estimates to.}
}
\description{
Bespoke function to create a \code{kable} from a \code{lm} object.
}
