% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-templates.R
\name{ggplot_counts_by_group}
\alias{ggplot_counts_by_group}
\title{Show Counts by Group}
\usage{
ggplot_counts_by_group(data, counts, group, ...)
}
\arguments{
\item{data}{tibble.}

\item{counts}{unquoted expression. Sub-level of grouping, mapped to \code{fill}.}

\item{group}{unquoted expression. Top-level of grouping, mapped to \code{x}.}

\item{...}{extra arguments passed to \code{\link[ggplot2]{geom_bar}}.}
}
\description{
Create a bar chart showing counts of 1 factor, grouped by another.
}
