% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-templates.R
\name{ggplot_mlm_estimates_by_year}
\alias{ggplot_mlm_estimates_by_year}
\title{Show Multilevel Estimates by Year}
\usage{
ggplot_mlm_estimates_by_year(tidy_tbl, alpha = 0.2, ...)
}
\arguments{
\item{tidy_tbl}{tibble. Output from \code{\link{tidy_fixef}} or
\code{\link{tidy_ranef}}.}

\item{..., alpha}{extra arguments passed to \code{\link[ggplot2]{geom_ribbon}}.}
}
\description{
Given a table of estimates will plot the estimates as a time series with
shaded region showing previously chosen credible interval.
}
