% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-templates.R
\name{ggplot_trend_by_group}
\alias{ggplot_trend_by_group}
\title{Show Trend of Time by Group}
\usage{
ggplot_trend_by_group(data, yvalue, group, nudge_x = 0, expand_x)
}
\arguments{
\item{data}{tibble.}

\item{yvalue}{unquoted expression. Column to plot on the y-axis.}

\item{group}{unquoted expression. Grouping column to colour lines.}

\item{nudge_x}{numeric. Value to move labels.}

\item{expand_x}{numeric. Expansion of x scale to improve label visibility.}
}
\description{
Time series plot with 1 line of distinct colour per group.
}
