% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rename_quantile_colnames}
\alias{rename_quantile_colnames}
\title{Replace Inconsistent Quantile Names}
\usage{
rename_quantile_colnames(x, repl = c("ci.low", "ci.high"))
}
\arguments{
\item{x}{object with valid \code{colnames}.}

\item{repl}{character vector, length 2. Replacement value, lowest first.}
}
\description{
Replace \code{colnames} such as \code{Q2.5} and \code{Q97.5} with tidy-like replacements.
}
