% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-brms.R
\name{tidy-brms-ef}
\alias{tidy-brms-ef}
\alias{tidy_fixef}
\alias{tidy_ranef}
\title{Tidy Fixed and Random Effect Estimates}
\usage{
tidy_fixef(x, ...)

tidy_ranef(x, ...)
}
\arguments{
\item{x}{\code{brms} fit.}

\item{...}{passed to corresponding \code{brms} method.}
}
\description{
Wrappers around population estimate getter'\code{\link[brms]{fixef.brmsfit}}
and group-level estimate getter \code{\link[brms]{ranef.brmsfit}}) that
tidies the data and decodes \verb{<chr>Year2019} to \verb{<int>2019}.
}
\section{Warning}{

Only use this method on \code{mlm-regional-epc} due to assumptions made.
}

